requirements.txt
streamlit==1.55.0
openai==2.30.0
python-dotenv==1.2.2
pywin32==311
pymupdf==1.27.2.2
pypdf==6.9.2
python-docx==1.2.0
openpyxl==3.1.5
reportlab==4.4.10
fastapi==0.117.1
uvicorn==0.36.0
python-multipart==0.0.20
